// src/app/models/menu/menu-titles.enum.ts
export enum AvailableSideMenuGroupTitles {
	Filter = "Filters",
	Effects = "Effects",
	Transform = "Transform"
}

export enum AvailableActionMenuGroupTitles {
	Action = "Action",
	Info = "Info"
}

export enum AvailableFiltersMenuItemTitles {
	None = "None",
	lofi = "loFi",
	radio = "radio",
	Oceanic = "oceanic",
	Marine = "marine",
	Liquid = "liquid",
	Diamante = "diamante",
	Twenties = "twenties",
	Mauve = "mauve",
	Vintage = "vintage"
}

export enum AvailableEffectsMenuItemTitles {
	None = "None",
	Solarize = "solarize",
	Oil_Painting = "oil painting",
	Offset_Red = "offset red",
	Ryo = "ryo"
}

export enum AvailableTransformMenuItemTitles {
	None = "None",
	Flip_Horizontal = "Flip Horizontal",
	Flip_Vertical = "Flip Vertical"
}

export enum AvailableActionMenuItemTitles {
	Download_Image = "Download Image",
	Clear_Image_Modifications = "Clear Image Modifications",
	Clear_Image = "Clear Image"
}